refactor: nest layout components in layout mod
