+++ src/site/mod.rs
@@ -4,12 +4,7 @@ use std::fs::File;
 use std::io::Write;
 use std::path::Path;
 pub mod artifacts;
-mod css;
-mod footer;
-mod head;
-mod header;
 pub mod html;
-mod javascript;
 mod layout;
 pub mod markdown;
 use crate::config::Config;
